requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
click==6.7
Flask==0.12.1
itsdangerous==0.24
Jinja2==2.9.5
lxml==3.7.3
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
pyspark==2.1.0+hadoop2.7
python-dotenv==0.6.4
qiniu==7.1.2
requests==2.13.0
six==1.10.0
tinydb==3.2.2
Werkzeug==0.12.1
